moved the client components to the leafs
