+++ app/utils/constants.ts
@@ -1,0 +1,3 @@
+export const TITLE_COLORS = ['#4a5968', '#a3999f', '#b4bdc5', '#f4d8cd'];
+
+// Path: app/utils/constants.ts
