+++ app/utils/types.ts
@@ -1,0 +1,4 @@
+export type ColorType = string[];
+
+// Path: app/utils/types.ts
+
